Webworker for recusiveDependencyFetch (#140)
* kara/webworker

1. Working test of using a webworker to recursively fetch dependencies.

* kara/webworker

Fix multiple re-renders and so spinning up of multiple webworkers.

* Move execution console.log into useEffect,

* kara/webworker

1. memoise dependencies and dependants.

1. add passive to message event handler

* kara/webworker

1. remove unecessary passive prop from event listener

1. add cache to dependencies for dependencies.

* webworker

1. Removed useRecursiveDependencyFetch.

1. Added useRecursiveDependencyFetch functionality to the main.js file.

* ESM hack.

* Merge in Luke's changes

* tidy up awaits

* recursive dependency fetching

* fix issue with switching files on long-running processes in webworkers.

Also remove recusrion param as we're always recursing.

* Fix worker.current bug

* Tidy up useEffect

* Fix merge conflicts
