+++ components/Main.js
@@ -1,7 +1,6 @@
 import { react, html, css } from 'rplus';
 import { useStateValue } from '../utils/globalState.js';
 import { parseUrl } from '../utils/parseUrl.js';
-import { parseDependencies } from '../utils/recursiveDependencyFetch.js';
 
 import Nav from './Nav.js';
 import Article from './Article.js';
@@ -66,12 +65,47 @@ export default () => {
   }, [request.name]);
 
   // Parse dependencies for the current code
+
+  const requestFileRef = react.useRef(null);
+  const worker = react.useRef(null);
+
+  // Set up listener for messages from the webworker
   react.useEffect(() => {
-    if (request.file)
-      parseDependencies('https://unpkg.com/' + request.path).then(cache =>
-        dispatch({ type: 'setCache', payload: cache })
-      );
-  }, [request.name, request.version, request.file]);
+    let buffer = {};
+    const eventListener = e => {
+      if (e.data.url === 'https://unpkg.com/' + request.path) {
+        dispatch({
+          type: 'setCache',
+          payload: { [e.data.url]: e.data },
+        });
+      } else {
+        buffer = { ...buffer, [e.data.url]: e.data };
+      }
+    };
+    if (request.file && requestFileRef.current !== request.file) {
+      requestFileRef.current = request.file;
+      worker.current = new Worker('./utils/recursiveDependencyFetchWorker.js');
+      worker.current.postMessage('https://unpkg.com/' + request.path);
+      if (worker.current) {
+        worker.current.addEventListener('message', eventListener);
+        setInterval(() => {
+          if (Object.keys(buffer).length > 0) {
+            dispatch({
+              type: 'setCache',
+              payload: buffer,
+            });
+            buffer = {};
+          }
+        }, 1000);
+      }
+    }
+    return () => {
+      if (worker.current) {
+        worker.current.terminate();
+        worker.current.removeEventListener('message', eventListener);
+      }
+    };
+  }, [request.path, request.file]);
 
   // Fetch packages by search term
   react.useEffect(() => {
