--- utils/makePath.js
@@ -1,30 +1,0 @@
-import fileNameRegEx from '../utils/fileNameRegEx.js';
-
-const UNPKG = 'https://unpkg.com/';
-
-// Handles paths like "../../some-file.js"
-const handleDoubleDot = (pathEnd, base) => {
-  const howFarBack = -1 * pathEnd.match(/\.\.\//g).length;
-  const strippedPathEnd = pathEnd.replace(/\.\./g, '').replace(/\/+/g, '/');
-  const strippedBase = base
-    .split('/')
-    .slice(0, howFarBack)
-    .join('/');
-  return strippedBase + strippedPathEnd;
-};
-
-// Handles cases like Svelte, where - on runpkg - the index url doesn't have a file ext
-const getCurrentdir = currentPath =>
-  currentPath.match(fileNameRegEx)
-    ? currentPath.replace(fileNameRegEx, '')
-    : currentPath.replace(/\/[^\/]+$/, '');
-
-const makePath = url => x => {
-  const base = getCurrentdir(url);
-  if (x.startsWith('./')) return base + x.replace('./', '/');
-  if (x.startsWith('../')) return handleDoubleDot(x, base);
-  if (x.startsWith('https://')) return x;
-  return UNPKG + x;
-};
-
-export default makePath;
